zmq4: optimize pub socket Send
This also changes the semantics of the pub socket to match that of czmq;
pubSocket.Send now returns immediately once the msg has been written
onto an outbound queue.

A new option, OptionHWM is added, per czmq semantics, to control the
high-water mark. Only the pub socket supports this option at the moment.