zmq4: implement proper REQ and REP socket
This fixes problems with REP sending messages to all connected peers
rather than to the originating peer.

The REQ/REP semantics now matches the recommended implementation behavior as specified in the ZMQ RFC.
In particular:
* REQ shall route outgoing messages to connected peers using a
round-robin strategy
* REQ shall block on sending, or return a suitable error, when it has no
connected peers.
* REP shall not block on sending

Thus, this changes the behavior a bit as REP no longer blocks on Send.

resolves #70