zmq4: fix router race/deadlock when multiple clients disconnect
This CL fixes a rare race condition that occurs when multiple clients
disconnect from a router socket in rapid succession, which results in a deadlock.

The lock under contention is routerMWriter.mu, which is taken in
routerMWriter.write() as part of a Socket.Send() call. What happens is that
the write fails and it tries to call socket.scheduleRmConn() to remove the failed
Conn (while still holding the lock). But scheduleRmConn() can't proceed because
the socket.closedConns channel is full, and the connReaper goroutine is stuck
trying to acquire the same lock via routerMWriter.rmConn.

The first commit here provides a test case that deadlocks reliably on my machine.
(The stack trace is messy because it includes noise from the client sockets, the
problem is much easier to see when the clients run as a separate process.)

The proposed fix just changes closedConns from a (fixed-capacity) channel into
an (arbitrary-sized) slice, managed by a sync.Cond.

Conceptually, instead of using an unbuffered channel for `closedConns`
(or a channel with any finite capacity), we use a mutex and condvar to
manage an arbitrary-sized list of closed connections.