Fix connection reaper deadlock.
Added explicitly running test to detect deadlock.
Use ```go test -tags=TestConnReaperDeadlockLoop -run ^TestConnReaperDeadlockLoop$```.
Be aware the test generate a lot of sockets in TIME_WAIT.
Without patch to socket.go it breaks pretty quick on fast and slow machines.
With the patch it runs way longer but gets slower when there are plenty of sockets in TIME_WAIT.